feat: adjustments
