+++ apps/antalmanac/src/lib/tabs/hooks.ts
@@ -1,16 +1,15 @@
-import { TAB_HREF, TAB_INDEX, isTabRouteSegment, type TabName } from '$lib/tabs/tabs';
+import { TAB_HREF, TAB_INDEX, type TabName } from '$lib/tabs/tabs';
 import { useCallback } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 
-/** Active tab index derived from the URL (`/`, `/calendar`, `/added`, `/map`). */
 export function useActiveTabIndex(): number {
     const { tab } = useParams();
 
-    if (isTabRouteSegment(tab)) {
-        return TAB_INDEX[tab];
+    if (!(tab === 'calendar') && !(tab === 'added') && !(tab === 'map')) {
+        return TAB_INDEX.search;
     }
 
-    return TAB_INDEX.search;
+    return TAB_INDEX[tab];
 }
 
 export function useGoToTab() {
